Changed the size of the address in binaries (now 32 bits)

+++ includes/transformation/irScheduler.h
@@ -6,7 +6,7 @@
 #ifndef __IR_SCHEDULER
 #define __IR_SCHEDULER
 
-int irScheduler(DBTPlateform *platform, uint1 optLevel, uint8 basicBlockSize, uint16 addressInBinaries,
+int irScheduler(DBTPlateform *platform, uint1 optLevel, uint8 basicBlockSize, uint32 addressInBinaries,
 		int6 numberFreeRegister, char configuration);
 
 #endif
